requirements.txt
fastapi==0.108.0
pydantic==2.5.3
python-dotenv==1.0.0
SQLAlchemy==2.0.25
alembic
setuptools
psycopg2-binary
python-jose 
passlib
bcrypt
loguru
pydantic[email]
